订单管理测试完毕

+++ src/main/java/org/imooc/service/CommentService.java
@@ -6,6 +6,6 @@ import org.imooc.dto.CommentListDto;
 public interface CommentService {
 
     //这个是实现查询评论信息
-    CommentListDto getListByBusinessId(Long businessId, Page page);
+    CommentListDto getListByBusinessId(int businessId, Page page);
 
 }
